copy paste with hidden cells.
